requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
